chore: integration test structure
