+++ hyperbench/integration_tests/__init__.py
@@ -1,0 +1,19 @@
+from .common import (
+    common_standard_metrics,
+    loaders,
+    model_configs,
+    multi_model_trainer,
+    splits_dataset,
+    datasets_enrichers,
+    add_negatives,
+)
+
+__all__ = [
+    "add_negatives",
+    "common_standard_metrics",
+    "datasets_enrichers",
+    "loaders",
+    "model_configs",
+    "multi_model_trainer",
+    "splits_dataset",
+]
