--- hyperbench/tests/integration/__init__.py
@@ -1,3 +1,0 @@
-from .common import gcn_model, hgnn_model, hgnnp_model
-
-__all__ = ["gcn_model", "hgnn_model", "hgnnp_model"]
